+++ src/main/java/it/mfm/model/ProductDao.java
@@ -42,7 +42,7 @@ public class ProductDao implements ProductDaoInterfaccia{
             preparedStatement = connection.prepareStatement(insertSQL);
             preparedStatement.setString(2, productBean.getNome());
             preparedStatement.setString(3, productBean.getDescrizione());
-            preparedStatement.setInt(4, productBean.getPrezzo());
+            preparedStatement.setDouble(4, productBean.getPrezzo());
             preparedStatement.setString(5, productBean.getImmagine());
             preparedStatement.setInt(6, productBean.getCategoria_id());
             preparedStatement.executeUpdate();
@@ -75,7 +75,7 @@ public class ProductDao implements ProductDaoInterfaccia{
             preparedStatement = connection.prepareStatement(updateSQL);
             preparedStatement.setString(1, productBean.getNome());
             preparedStatement.setString(2, productBean.getDescrizione());
-            preparedStatement.setInt(3, productBean.getPrezzo());
+            preparedStatement.setDouble(3, productBean.getPrezzo());
             preparedStatement.setString(4, productBean.getImmagine());
             preparedStatement.setInt(5, productBean.getCategoria_id());
             preparedStatement.setInt(6, productBean.getId());
@@ -143,7 +143,7 @@ public class ProductDao implements ProductDaoInterfaccia{
                 productBean.setId(resultSet.getInt("id"));
                 productBean.setNome(resultSet.getString("nome"));
                 productBean.setDescrizione(resultSet.getString("descrizione"));
-                productBean.setPrezzo(resultSet.getInt("prezzo"));
+                productBean.setPrezzo(resultSet.getDouble("prezzo"));
                 productBean.setImmagine(resultSet.getString("immagine"));
                 productBean.setCategoria_id(resultSet.getInt("categoria_id"));
             }
@@ -184,7 +184,7 @@ public class ProductDao implements ProductDaoInterfaccia{
                 productBean.setId(resultSet.getInt("id"));
                 productBean.setNome(resultSet.getString("nome"));
                 productBean.setDescrizione(resultSet.getString("descrizione"));
-                productBean.setPrezzo(resultSet.getInt("prezzo"));
+                productBean.setPrezzo(resultSet.getDouble("prezzo"));
                 productBean.setImmagine(resultSet.getString("immagine"));
                 productBean.setCategoria_id(resultSet.getInt("categoria_id"));
                 productList.add(productBean);
@@ -227,7 +227,7 @@ public class ProductDao implements ProductDaoInterfaccia{
                 productBean.setId(resultSet.getInt("id"));
                 productBean.setNome(resultSet.getString("nome"));
                 productBean.setDescrizione(resultSet.getString("descrizione"));
-                productBean.setPrezzo(resultSet.getInt("prezzo"));
+                productBean.setPrezzo(resultSet.getDouble("prezzo"));
                 productBean.setImmagine(resultSet.getString("immagine"));
                 productBean.setCategoria_id(resultSet.getInt("categoria_id"));
                 productList.add(productBean);
